chore: upgrade claude-agent-sdk from 0.1.44 to 0.1.45
- Bundled CLI updated to v2.1.63
- Fix witness RESTART action: cancel_run → await runner.cancel()
- Fix doctor stale runs null guard on started_at
- Fix merge queue backoff return type (int cast)
- Update test mock to match corrected cancel method

+++ src/gluon/merge_queue.py
@@ -152,4 +152,4 @@ class MergeQueueService:
 
     def calculate_backoff(self, conflict_count: int) -> int:
         """Exponential backoff: 60 * 2^count seconds."""
-        return 60 * (2**conflict_count)
+        return int(60 * (2**conflict_count))
